+++ entrypoint.sh
@@ -5,7 +5,7 @@ source ${REDMINE_RUNTIME_DIR}/functions
 [[ $DEBUG == true ]] && set -x
 
 case ${1} in
-  app:init|app:start|app:rake|app:backup:create)
+  app:init|app:start|app:rake|app:backup:create|app:backup:restore)
 
     initialize_system
     configure_redmine
@@ -36,6 +36,10 @@ case ${1} in
         shift 1
         backup_create $@
         ;;
+      app:backup:restore)
+        shift 1
+        backup_restore $@
+        ;;
     esac
     ;;
   app:help)
@@ -44,6 +48,7 @@ case ${1} in
     echo " app:init           - Initialize the Redmine server (e.g. create databases, install plugins/themes), but don't start it."
     echo " app:rake <task>    - Execute a rake task."
     echo " app:backup:create  - Create a backup."
+    echo " app:backup:restore - Restore an existing backup."
     echo " app:help           - Displays the help"
     echo " [command]          - Execute the specified command, eg. bash."
     ;;